+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/lsfiles/compare/v0.2.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+-  Add `LSFiles.populate_regex`
 
 [0.2.0](https://github.com/jshwi/lsfiles/releases/tag/v0.2.0) - 2023-01-03
 ------------------------------------------------------------------------